+++ src/collective/contact/contactlist/adapter.py
@@ -1,8 +1,12 @@
-from zope.interface import Interface, implements
+from zc.relation.interfaces import ICatalog
+from zope.component import getUtility
 from zope.component import adapts
+from zope.interface import Interface, implements
+from zope.intid.interfaces import IIntIds
+
 from plone import api as ploneapi
-from collective.contact.contactlist.interfaces import IUserLists
-from collective.contact.contactlist.content.contactlist import IContactList
+
+from collective.contact.contactlist.interfaces import IUserLists, IContactList
 
 
 class  UserListStorage(object):
@@ -47,3 +51,13 @@ class  UserListStorage(object):
             return None
 
         return self.portal.Members.get(user_id, None)
+
+    def get_lists_for_contact(self, contact):
+        """Get lists that contain contact."""
+        catalog = getUtility(ICatalog)
+        intids = getUtility(IIntIds)
+        contact_intid = intids.queryId(contact)
+        query = {'to_id': contact_intid,
+                 'from_interfaces_flattened': IContactList,
+                 'from_attribute': 'contacts'}
+        return [rel.from_object for rel in catalog.findRelations(query)]