+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 1.1 (unreleased)
 ----------------
 
+- Add get_lists_for_item method to get all lists that contain contact.
+  [cedricmessiant]
+
 - Fix relations.
   [cedricmessiant]